validate data and add department entity

+++ Dubox.Application/Features/Users/Commands/AssignRolesToUserCommandHandler.cs
@@ -1,6 +1,6 @@
+using Dubox.Domain.Abstraction;
 using Dubox.Domain.Entities;
 using Dubox.Domain.Shared;
-using Dubox.Domain.Abstraction;
 using MediatR;
 
 namespace Dubox.Application.Features.Users.Commands;
@@ -22,10 +22,18 @@ public class AssignRolesToUserCommandHandler : IRequestHandler<AssignRolesToUser
         if (user == null)
             return Result.Failure("User not found");
 
+        var existingRolesCount = _unitOfWork.Repository<Role>()
+       .Get().Count(r => request.RoleIds.Contains(r.RoleId));
+
+        if (existingRolesCount != request.RoleIds.Count)
+        {
+            return Result.Failure("One or more roles were not found in the roles.");
+        }
         var existingUserRoles = _unitOfWork.Repository<UserRole>()
             .Get()
             .Where(ur => ur.UserId == request.UserId)
             .ToList();
+
 
         _unitOfWork.Repository<UserRole>().DeleteRange(existingUserRoles);
